% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lrt.R
\name{bootstrap_lrt}
\alias{bootstrap_lrt}
\title{Apply a likelihod ratio test}
\usage{
bootstrap_lrt(
  W,
  fitted_model,
  null_param,
  n_boot,
  m = NULL,
  recalculate_W0 = FALSE,
  parallelize = FALSE,
  ncores = 1,
  save_models = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{W}{An \eqn{n \times J} matrix of numeric HTS output (e.g., read counts, coverages, etc.)}

\item{fitted_model}{The model fitted under the alternative}

\item{null_param}{The model fitted under the null. This could be a modified copy of the model fit under the alternative.}

\item{n_boot}{The number of bootstrap resamples to perform}

\item{m}{description}

\item{recalculate_W0}{Boolean. Should rescaling W. TODO. Defaults to FALSE.}

\item{parallelize}{Boolean. do you want to parallelize it?}

\item{ncores}{the number of cores to parallelize over}

\item{save_models}{TODO}

\item{verbose}{Do you want to know what I'm doing? Defaults to FALSE.}
}
\description{
Apply a likelihod ratio test
}
