% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_log_penalty.R
\name{calculate_log_penalty}
\alias{calculate_log_penalty}
\title{Calculate barrier penalty to add to objective function inside
barrier algorithm}
\usage{
calculate_log_penalty(varying_lr_df, fixed_df, barrier_t)
}
\arguments{
\item{varying_lr_df}{A data frame containing values of parameters that
are treated as unknown, with relative abundance parameters represented
on the log ratio scale (i.e., as phi and phi_tilde)}

\item{fixed_df}{A data frame containing values of parameters that are
treated as known}

\item{barrier_t}{The current value of t, the barrier penalty parameter}
}
\value{
The calculated value of the barrier penalty
}
\description{
Calculate barrier penalty to add to objective function inside
barrier algorithm
}
