% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{costea2017_metaphlan2_profiles}
\alias{costea2017_metaphlan2_profiles}
\title{MetaPhlAn2 taxonomic profiles from the Costea 2017 study.}
\format{
A data frame with the observed relative abundances (from MetaPhlAn2) of taxa (rows) in each sample (columns)
}
\source{
\url{https://doi.org/10.1038/nbt.3960}
}
\usage{
costea2017_metaphlan2_profiles
}
\description{
MetaPhlAn2 taxonomic profiles from the Costea 2017 study.
}
\keyword{datasets}
