% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{costea2017_mock_composition}
\alias{costea2017_mock_composition}
\title{The composition of 10 spike-in taxa in 19 samples observed from flow cytometry.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 6 columns.
}
\source{
\url{https://doi.org/10.1038/nbt.3960}
}
\usage{
costea2017_mock_composition
}
\description{
For details, see Costea et al (2017).
}
\keyword{datasets}
