% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes_to_parameters.R
\name{dataframes_to_parameters}
\alias{dataframes_to_parameters}
\title{Convert parameter values stored in data frame format to matrix format}
\usage{
dataframes_to_parameters(fixed_df, varying_df)
}
\arguments{
\item{fixed_df}{A dataframe containing values of model parameters
treated as fixed and known (i.e. held constant at known values)}

\item{varying_df}{A dataframe containing current values of model parameters treated
as fixed and unknown (i.e., parameters to be estimated)}
}
\value{
A list containing
\item{P}{Specimen relative abundance matrix (of dimension K x J)}
\item{P_tilde}{Spurious read source
relative abundance matrix (of dimension K-tilde x J)}
\item{B}{A matrix of detection efficiencies (of dimension p x J)}
\item{gammas}{An n-vector of sample-specific read intensities}
\item{gamma_tilde}{A-vector of spurious read source intensities
(of length K-tilde)}
}
\description{
Convert parameter values stored in data frame format to matrix format
}
\author{
David Clausen
}
