% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Fit tinyvamp model to HTS microbiome data}
\usage{
estimate_parameters(
  W,
  X,
  Z,
  Z_tilde = NULL,
  Z_tilde_gamma_cols,
  gammas,
  gammas_fixed_indices,
  P,
  P_fixed_indices,
  B,
  B_fixed_indices,
  X_tilde,
  P_tilde,
  P_tilde_fixed_indices,
  gamma_tilde,
  gamma_tilde_fixed_indices,
  alpha_tilde = NULL,
  Z_tilde_list = NULL,
  barrier_t = 1,
  barrier_scale = 10,
  max_barrier = 1e+12,
  initial_conv_tol = 1000,
  final_conv_tol = 0.1,
  final_f = 1e-06,
  constraint_tolerance = 1e-10,
  hessian_regularization = 0.01,
  criterion = "Poisson",
  subproblem_method = "Newton",
  profile_P = TRUE,
  barrier_maxit = 500,
  profiling_maxit = 25,
  wts = NULL,
  verbose = FALSE,
  bootstrap_failure_cutoff = NULL,
  return_variance = FALSE
)
}
\arguments{
\item{W}{An \eqn{n \times J} matrix of numeric HTS output (e.g., read counts, coverages, etc.)}

\item{X}{The sample efficiency design -- an \eqn{n \times p} matrix}

\item{Z}{The sample-specimen design -- an \eqn{n \times K} matrix whose \eqn{ij}-th entry
indicates the proportional contribution of specimen \eqn{j} to sample \eqn{i}. Rows must
sum to 1 or be identically 0.}

\item{Z_tilde}{The spurious read design -- an \eqn{n x \tilde{K}} matrix where
\eqn{\tilde{K}} is the number of spurious read sources modeled.}

\item{Z_tilde_gamma_cols}{A numeric vector containing the columns of Z_tilde which should be
multiplied by exp(gamma).}

\item{gammas}{A numeric vector of length n of starting values for read intensity parameter gamma}

\item{gammas_fixed_indices}{A logical vector of length n whose \eqn{i}-th entry is TRUE if the
\eqn{i}-th entry of gamma should be treated as fixed and known, and FALSE otherwise}

\item{barrier_t}{, #starting value of reciprocal barrier penalty coef.}

\item{barrier_scale}{= 10, #increments for value of barrier penalty}

\item{max_barrier}{= 1e20, #maximum value of barrier_t}

\item{final_f}{= 1e-6,}

\item{constraint_tolerance}{= 1e-15,}

\item{hessian_regularization}{= 0.01,}

\item{subproblem_method}{= "Newton"}

\item{profile_P}{= TRUE Run profiling step after barrier algorithm
has run? If TRUE, this step is performed, possibly setting some estimated
relative abundances in P equal to zero. If FALSE, profiling step is skipped and
back-transformed log-ratio parameter estimated via barrier algorithm is
returned for P.
@param profiling_maxit Maximum number of iterations to run profiling step
in P for (default is 25).}

\item{final_epsilon}{= 0.01,}
}
\value{
A list containing
\item{pval}{The p-value}
}
\description{
This function fits a model to HTS microbiome data that allows for estimation of
detection efficiency effects as well as modeling of spurious read sources
(e.g., contamination).
}
\author{
David Clausen
}
