% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{karstens_phyloseq}
\alias{karstens_phyloseq}
\title{karstens_phyloseq}
\format{
A \code{phyloseq} object containing otu_table, tax_table, and sample_data.
}
\source{
\url{https://doi.org/10.1128/mSystems.00290-19}
}
\usage{
karstens_phyloseq
}
\description{
A phyloseq object used in package examples and analyses.
}
\keyword{datasets}
