% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_jac_lr_faster.R
\name{mean_jac_lr_faster}
\alias{mean_jac_lr_faster}
\title{Fit tinyvamp model to HTS microbiome data}
\usage{
mean_jac_lr_faster(
  fixed_df,
  varying_lr_df,
  varying_df,
  which_k_p,
  which_k_p_tilde,
  which_B_rows,
  which_B_keep,
  which_gammas,
  which_gamma_tilde,
  params,
  Ak_list,
  A_tilde_k_list,
  fixed_P_multipliers,
  fixed_P_tilde_multipliers,
  K,
  K_tilde,
  X,
  Z,
  X_tilde,
  Z_tilde,
  Z_tilde_gamma_cols,
  Z_tilde_list = NULL,
  sparse = TRUE,
  proportion_scale = FALSE,
  P_fixed_indices = NULL,
  P_tilde_fixed_indices = NULL
)
}
\arguments{
\item{fixed_df}{A dataframe containing values of model parameters
treated as fixed and known (i.e. held constant at known values)}

\item{varying_df}{A dataframe containing current values of model parameters treated
as fixed and unknown (i.e., parameters to be estimated)}

\item{X}{The sample efficiency design -- an \eqn{n \times p} matrix}

\item{Z}{The sample-specimen design -- an \eqn{n \times K} matrix whose \eqn{ij}-th entry
indicates the proportional contribution of specimen \eqn{j} to sample \eqn{i}. Rows must
sum to 1 or be identically 0.}

\item{Z_tilde}{The spurious read design -- an \eqn{n x \tilde{K}} matrix where
\eqn{\tilde{K}} is the number of spurious read sources modeled.}

\item{Z_tilde_gamma_cols}{A numeric vector containing the columns of Z_tilde which should be
multiplied by exp(gamma).}

\item{sparse}{Use sparsity in Jacobian to speed up computation (default is TRUE)}
}
\value{
A list containing
\item{pval}{The p-value}
}
\description{
This function fits a model to HTS microbiome data that allows for estimation of
detection efficiency effects as well as modeling of spurious read sources
(e.g., contamination).
}
\author{
David Clausen
}
