% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaninate.R
\name{meaninate}
\alias{meaninate}
\title{Calculate the mean given parameters}
\usage{
meaninate(
  gammas,
  B,
  X,
  Z,
  P,
  X_tilde,
  Z_tilde = NULL,
  Z_tilde_gamma_cols,
  P_tilde,
  gamma_tilde,
  alpha_tilde = NULL,
  Z_tilde_list = NULL,
  return_separate = FALSE,
  exclude_gammas = FALSE
)
}
\arguments{
\item{gammas}{A numeric vector of length n of starting values for read intensity parameter gamma}

\item{B}{A \eqn{p \times J} numeric matrix giving initial values for the sample efficiencies.}

\item{X}{The sample efficiency design -- an \eqn{n \times p} matrix}

\item{Z}{The sample-specimen design -- an \eqn{n \times K} matrix whose \eqn{ij}-th entry
indicates the proportional contribution of specimen \eqn{j} to sample \eqn{i}. Rows must
sum to 1 or be identically 0.}

\item{P}{A \eqn{K \times J} numeric matrix giving initial values for the relative abundance matrix.}

\item{X_tilde}{A \eqn{\tilde{K} \times p} matrix giving the spurious read source efficiency design matrix}

\item{Z_tilde}{The spurious read design -- an \eqn{n \times \tilde{K}} matrix where
\eqn{\tilde{K}} is the number of spurious read sources modeled.}

\item{Z_tilde_gamma_cols}{A numeric vector containing the columns of Z_tilde which should be
multiplied by exp(gamma).}

\item{P_tilde}{A \eqn{\tilde{K} \times J} numeric matrix giving initial values for the spurious read source relative abundances.}

\item{gamma_tilde}{A numeric vector of length \eqn{\tilde{K}} of starting values for spurious read intensity parameter gamma_tilde}

\item{alpha_tilde}{A numeric vector containing starting values of length \eqn{M}. If used, \code{Z_tilde_list} must be provided.}

\item{Z_tilde_list}{A list of length \eqn{M + 1} containing matrices \eqn{\tilde{Z}_1,\dots,\tilde{Z}_{M + 1}} to be linearly combined to
generate \code{Z_tilde}: \eqn{\tilde{Z} = \tilde{Z}_{(1)} + \sum_{m = 1}^M \exp(\tilde{\alpha}_m)\tilde{Z}_{(m + 1)}}. If used,
\code{alpha_tilde} must be provided.}

\item{return_separate}{Boolean. Return the summed mean, or separate the sample and contamination pieces. Defaults to FALSE.}

\item{exclude_gammas}{Boolean, defaults to FALSE. Should the gamma components be ignored?}
}
\description{
Calculate the mean given parameters
}
