% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_d_gamma_tilde.R
\name{mu_d_gamma_tilde}
\alias{mu_d_gamma_tilde}
\title{Calculate derivative of mu_ij with respect to an entry of gamma_tilde}
\usage{
mu_d_gamma_tilde(
  i,
  j,
  k_tilde,
  gammas,
  B,
  X,
  Z,
  P,
  X_tilde,
  Z_tilde,
  Z_tilde_gamma_cols,
  alpha_tilde = NULL,
  Z_tilde_list = NULL,
  P_tilde,
  gamma_tilde
)
}
\arguments{
\item{i}{The sample index (must be in 1, ..., n)}

\item{j}{The taxon index (must be in 1, ..., J)}

\item{k_tilde}{The element of gamma_tilde with respect to which to take derivative}

\item{gammas}{Numeric vector of read intensities}

\item{B}{Detection efficiency matrix}

\item{X}{The efficiency design matrix (n x p)}

\item{Z}{The sample design matrix (n x K)}

\item{P}{The sample relative abundance matrix (K x J)}

\item{X_tilde}{The spurious read efficiency design (K_tilde x p)}

\item{Z_tilde}{The spurious read design (n x K_tilde)}

\item{Z_tilde_gamma_cols}{Numeric vector containing indexes of columns of
Z_tilde to scale by exp(gamma); NULL if no columns to be scaled}

\item{P_tilde}{The spurious source relative abundance matrix (K_tilde x J)}

\item{gamma_tilde}{Spurious read intensity parameter}
}
\value{
A derivative d mu_ij / d gamma_tilde_k_tilde
}
\description{
Calculate derivative of mu_ij with respect to an entry of gamma_tilde
}
