% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_d_P.R
\name{mu_d_P}
\alias{mu_d_P}
\title{Calculate derivative of mu_ij with respect to an entry of P}
\usage{
mu_d_P(i, j, m, gammas, B, X, Z, P)
}
\arguments{
\item{i}{The sample index (must be in 1, ..., n)}

\item{j}{The taxon index (must be in 1, ..., J)}

\item{m}{The row of P with respect to which to take derivative}

\item{gammas}{Numeric vector of read intensities}

\item{B}{Detection efficiency matrix}

\item{X}{The efficiency design matrix (n x p)}

\item{Z}{The sample design matrix (n x K)}

\item{P}{The sample relative abundance matrix (K x J)}
}
\value{
A derivative d mu_ij / d P_kj
}
\description{
Calculate derivative of mu_ij with respect to an entry of P
}
