% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_mu_d_rho.R
\name{mu_d_rho_faster}
\alias{mu_d_rho_faster}
\title{Calculate derivative of mu_ij with respect to a row of matrix-valued
parameter rho}
\usage{
mu_d_rho_faster(i, J, k, gammas, B, X, Z, Ak_list, rho_k, fixed_P_multipliers)
}
\arguments{
\item{i}{The sample index (must be in 1, ..., n)}

\item{J}{The total number of taxa modeled}

\item{k}{Row index (which row of rho with respect to which to take derivative)}

\item{gammas}{Numeric vector of read intensities}

\item{B}{Detection efficiency matrix}

\item{X}{The efficiency design matrix (n x p)}

\item{Z}{The sample design matrix (n x K)}

\item{Ak_list}{List containing matrices that map back-transformed
rho to entries of P}

\item{rho_k}{Value of kth row of rho}

\item{fixed_P_multipliers}{Numeric vector of length K containing values in (0,1]
equal to 1 - sum(fixed relative abundances in row k of P)
Z_tilde to scale by exp(gamma); NULL if no columns to be scaled}
}
\value{
A derivative d mu_i / d rho_k
}
\description{
Calculate derivative of mu_ij with respect to a row of matrix-valued
parameter rho
}
