% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_d_rho_tilde_faster.R
\name{mu_d_rho_tilde_faster}
\alias{mu_d_rho_tilde_faster}
\title{Calculate derivative of mu_i with respect to a row of matrix-valued
parameter rho_tilde.}
\usage{
mu_d_rho_tilde_faster(
  i,
  J,
  k_tilde,
  gammas,
  B,
  rho_tilde_k,
  A_tilde_k_list,
  fixed_P_tilde_multipliers,
  X_tilde,
  Z_tilde,
  Z_tilde_gamma_cols,
  alpha_tilde = NULL,
  Z_tilde_list = NULL,
  gamma_tilde,
  proportion_scale = FALSE
)
}
\arguments{
\item{i}{The sample index (must be in 1, ..., n)}

\item{J}{The total number of taxa modeled}

\item{k_tilde}{Row index (which row of rho_tilde with respect to which to take derivative)}

\item{gammas}{Numeric vector of read intensities}

\item{B}{Detection efficiency matrix}

\item{rho_tilde_k}{Value of kth row of rho_tilde}

\item{A_tilde_k_list}{List containing matrices that map back-transformed
rho_tilde to entries of P_tilde}

\item{fixed_P_tilde_multipliers}{Numeric vector of row-specific multipliers
applied to the derivative of the transformation from
\eqn{\tilde{\rho}} to \eqn{\tilde{P}}.}

\item{X_tilde}{The spurious read efficiency design (K_tilde x p)}

\item{Z_tilde}{The spurious read design (n x K_tilde)}

\item{Z_tilde_gamma_cols}{Numeric vector containing indexes of columns of
Z_tilde to scale by exp(gamma); NULL if no columns to be scaled}

\item{alpha_tilde}{A numeric vector containing starting values of length \eqn{M}. If used, \code{Z_tilde_list} must be provided.}

\item{Z_tilde_list}{A list of length \eqn{M + 1} containing matrices \eqn{\tilde{Z}_1,\dots,\tilde{Z}_{M + 1}} to be linearly combined to
generate \code{Z_tilde}: \eqn{\tilde{Z} = \tilde{Z}_{(1)} + \sum_{m = 1}^M \exp(\tilde{\alpha}_m)\tilde{Z}_{(m + 1)}}. If used,
\code{alpha_tilde} must be provided.}

\item{gamma_tilde}{Numeric vector of row-specific intercept terms for the
spurious read model.}

\item{proportion_scale}{Logical. If \code{FALSE}, return the derivative of
\eqn{\mu_i} with respect to row \code{k_tilde} of \eqn{\tilde{\rho}}. If
\code{TRUE}, return only the derivative with respect to the corresponding
row of \eqn{\tilde{P}}.}
}
\value{
A derivative d mu_i / d rho_tilde_k
}
\description{
Calculate derivative of mu_i with respect to a row of matrix-valued
parameter rho_tilde.
}
