% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ci.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Apply the Bayesian subsampled bootstrap to a fitted tinyvamp model}
\usage{
bootstrap_ci(
  W,
  fitted_model,
  n_boot,
  m = NULL,
  alpha = 0.05,
  parallelize = FALSE,
  ncores = 5,
  seed = NULL,
  return_models = FALSE,
  verbose = FALSE,
  adjust = FALSE
)
}
\description{
Apply the Bayesian subsampled bootstrap to a fitted tinyvamp model
}
