% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Fit tinyvamp model to HTS microbiome data}
\usage{
estimate_parameters(
  W,
  X,
  Z,
  Z_tilde = NULL,
  Z_tilde_gamma_cols,
  gammas,
  gammas_fixed_indices,
  P,
  P_fixed_indices,
  B,
  B_fixed_indices,
  X_tilde,
  P_tilde,
  P_tilde_fixed_indices,
  gamma_tilde,
  gamma_tilde_fixed_indices,
  alpha_tilde = NULL,
  Z_tilde_list = NULL,
  barrier_t = 1,
  barrier_scale = 10,
  max_barrier = 1e+12,
  initial_conv_tol = 1000,
  final_conv_tol = 0.1,
  constraint_tolerance = 1e-10,
  hessian_regularization = 0.01,
  criterion = "Poisson",
  profile_P = TRUE,
  barrier_maxit = 500,
  profiling_maxit = 25,
  wts = NULL,
  verbose = FALSE,
  bootstrap_failure_cutoff = NULL,
  tinker_zeroes = 0.1,
  return_variance = FALSE
)
}
\arguments{
\item{W}{An \eqn{n \times J} matrix of numeric HTS output (e.g., read counts, coverages, etc.)}

\item{X}{The sample efficiency design -- an \eqn{n \times p} matrix}

\item{Z}{The sample-specimen design -- an \eqn{n \times K} matrix whose \eqn{ij}-th entry
indicates the proportional contribution of specimen \eqn{j} to sample \eqn{i}. Rows must
sum to 1 or be identically 0.}

\item{Z_tilde}{The spurious read design -- an \eqn{n \times \tilde{K}} matrix where
\eqn{\tilde{K}} is the number of spurious read sources modeled.}

\item{Z_tilde_gamma_cols}{A numeric vector containing the columns of Z_tilde which should be
multiplied by exp(gamma).}

\item{gammas}{A numeric vector of length n of starting values for read intensity parameter gamma}

\item{gammas_fixed_indices}{A logical vector of length n whose \eqn{i}-th entry is TRUE if the
\eqn{i}-th entry of gamma should be treated as fixed and known, and FALSE otherwise}

\item{P}{A \eqn{K \times J} numeric matrix giving initial values for the relative abundance matrix.}

\item{P_fixed_indices}{P_fixed_indices A \eqn{K \times J} logical matrix specifying any entries of P that are known. If known, the corresponding values from \code{P} will be treated as the fixed, known values.}

\item{B}{A \eqn{p \times J} numeric matrix giving initial values for the sample efficiencies.}

\item{B_fixed_indices}{A \eqn{p \times J} logical matrix specifying any entries of B that are known. If known, the corresponding values from \code{B} will be treated as the fixed, known values.}

\item{X_tilde}{A \eqn{\tilde{K} \times p} matrix giving the spurious read source efficiency design matrix}

\item{P_tilde}{A \eqn{\tilde{K} \times J} numeric matrix giving initial values for the spurious read source relative abundances.}

\item{P_tilde_fixed_indices}{A \eqn{\tilde{K} \times J} logical matrix indicating if the \eqn{(i,j)}th entry of \code{P_tilde} should be treated as fixed and known.}

\item{gamma_tilde}{A numeric vector of length \eqn{\tilde{K}} of starting values for spurious read intensity parameter gamma_tilde}

\item{gamma_tilde_fixed_indices}{A logical vector of length \eqn{\tilde{K}} whose \eqn{i}-th entry is TRUE if the
\eqn{i}-th entry of gamma_tilde should be treated as fixed and known, and FALSE otherwise.}

\item{alpha_tilde}{A numeric vector containing starting values of length \eqn{M}. If used, \code{Z_tilde_list} must be provided.}

\item{Z_tilde_list}{A list of length \eqn{M + 1} containing matrices \eqn{\tilde{Z}_1,\dots,\tilde{Z}_{M + 1}} to be linearly combined to
generate \code{Z_tilde}: \eqn{\tilde{Z} = \tilde{Z}_{(1)} + \sum_{m = 1}^M \exp(\tilde{\alpha}_m)\tilde{Z}_{(m + 1)}}. If used,
\code{alpha_tilde} must be provided.}

\item{barrier_t}{Starting value of reciprocal barrier penalty coef. Defaults to 1.}

\item{barrier_scale}{Increments for value of barrier penalty. Defaults to 10.}

\item{max_barrier}{Maximum value of barrier_t. Defaults to 1e12.}

\item{constraint_tolerance}{The tolerance for the augmented Lagrangian algorithm. Final estimates of P are relative abundances to within \code{constraint_tolerance} of 1, i.e., abs(sum p_{kj} - 1) <  \code{constraint_tolerance}. Defaults to 1e-10.}

\item{hessian_regularization}{The second step of optimization involves a quadratic approximation to the likelihood, for which we use a modified Taylor series for stability. This is the constant that dampens the second term. Defaults to 0.01.}

\item{criterion}{Should the algorithm return the Poisson maximum likelihood estimates or the reweighted Poisson maximum likelihood estimates? Options are "Poisson" or "reweighted_Poisson".}

\item{profile_P}{Defaults to TRUE Run profiling step after barrier algorithm has run? If TRUE, this step is performed, possibly setting some estimated relative abundances in P equal to zero. If FALSE, profiling step is skipped and back-transformed log-ratio parameter estimated via barrier algorithm is returned for P.}

\item{barrier_maxit}{The maximum number of iterations for the barrier method}

\item{profiling_maxit}{Maximum number of iterations to run profiling step in P for (default is 25).}

\item{wts}{Weights for reweighting the likelihood contributions. This is usually done to improve efficiency. Defaults to NULL. We compute the weights for you even if you choose \code{criterion = "reweighted_Poisson"}.}

\item{verbose}{Do you want to know what I'm doing? Defaults to FALSE.}

\item{bootstrap_failure_cutoff}{Defaults to NULL.}

\item{tinker_zeroes}{Because the barrier method can only be applied to relative abundances in the interior of the simplex, tinker_zeroes divided by the number of taxa is added to all relative abundances to be estimated before the barrier method is applied. Default 0.1.}

\item{return_variance}{Defaults to FALSE.}
}
\value{
A list containing estimated parameter values, along with the given inputs
}
\description{
This function fits a model to HTS microbiome data that allows for estimation of
detection efficiency effects as well as modeling of spurious read sources
(e.g., contamination).
}
\author{
David Clausen

Amy Willis
}
