% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_criterion_lr.R
\name{evaluate_criterion_lr}
\alias{evaluate_criterion_lr}
\title{Criterion Evaluation Function}
\usage{
evaluate_criterion_lr(
  W,
  X,
  Z,
  Z_tilde,
  Z_tilde_gamma_cols,
  Z_tilde_list = NULL,
  X_tilde,
  fixed_df,
  varying_df,
  varying_lr_df = NULL,
  barrier_t = NULL,
  criterion = "Poisson",
  lr_scale = TRUE,
  include_log_penalty = TRUE,
  wts = NULL,
  gmm_inv_wts = NULL,
  return_gmm_inv_weights = FALSE
)
}
\description{
Criterion Evaluation Function
}
