% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_mu_d_P_original.R
\name{mu_d_P}
\alias{mu_d_P}
\title{Calculate derivative of mu_ij with respect to an entry of P}
\usage{
mu_d_P(i, j, m, gammas, B, X, Z, P)
}
\arguments{
\item{i}{The sample index (must be in 1, ..., n)}

\item{j}{The taxon index (must be in 1, ..., J)}

\item{m}{The row of P with respect to which to take derivative}

\item{gammas}{Numeric vector of read intensities}

\item{B}{Detection efficiency matrix}

\item{X}{The efficiency design matrix (n x p)}

\item{Z}{The sample design matrix (n x K)}

\item{P}{The sample relative abundance matrix (K x J)}

\item{X_tilde}{The spurious read efficiency design (K_tilde x p)}

\item{Z_tilde}{The spurious read design (n x K_tilde)}

\item{Z_tilde_gamma_cols}{Numeric vector containing indexes of columns of
Z_tilde to scale by exp(gamma); NULL if no columns to be scaled}

\item{P_tilde}{The spurious source relative abundance matrix (K_tilde x J)}

\item{gamma_tilde}{Spurious read intensity parameter}
}
\value{
A derivative d mu_ij / d P_kj
}
\description{
Calculate derivative of mu_ij with respect to an entry of P
}
