% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsum.R
\name{sum_of_logs}
\alias{sum_of_logs}
\title{Sum positive numbers via their logarithms
Given x and y, this function returns log(exp(x) + exp(y)).}
\usage{
sum_of_logs(x, y = NULL)
}
\arguments{
\item{x}{the thing to be summed.}

\item{y}{the thing to be added to x. If absent, just the sum over vector-valued x is taken.}
}
\description{
Sum positive numbers via their logarithms
Given x and y, this function returns log(exp(x) + exp(y)).
}
\author{
David Clausen
}
